created a custom hook for number format translation
