+++ src/common/Cover/CoverCard.jsx
@@ -2,7 +2,6 @@ import { Divider } from "@/common/Divider/Divider";
 import { ProgressBar } from "@/common/ProgressBar/ProgressBar";
 import { OutlinedCard } from "@/common/OutlinedCard/OutlinedCard";
 import { getCoverImgSrc } from "@/src/helpers/cover";
-import { formatCurrency } from "@/utils/formatter/currency";
 import { convertFromUnits, toBN } from "@/utils/bn";
 import { formatPercent } from "@/utils/formatter/percent";
 import { MULTIPLIER } from "@/src/config/constants";
@@ -12,6 +11,7 @@ import { CardStatusBadge } from "@/common/CardStatusBadge";
 import { useMyLiquidityInfo } from "@/src/hooks/provide-liquidity/useMyLiquidityInfo";
 import { useCommitment } from "@/src/hooks/provide-liquidity/useCommitment";
 import { Trans } from "@lingui/macro";
+import { useNumberFormat } from "@/src/hooks/useNumberFormat";
 
 export const CoverCard = ({ details }) => {
   const { projectName, key, ipfsData } = details;
@@ -20,6 +20,7 @@ export const CoverCard = ({ details }) => {
   const { info: liquidityInfo } = useMyLiquidityInfo({ coverKey: key });
   const { commitment } = useCommitment({ coverKey: key });
   const imgSrc = getCoverImgSrc(coverInfo);
+  const { formatCurrency } = useNumberFormat();
 
   const liquidity = liquidityInfo.totalLiquidity;
   const protection = commitment;
