+++ src/common/PolicyFeesAndExpiry/PolicyFeesAndExpiry.jsx
@@ -1,11 +1,12 @@
 import DateLib from "@/lib/date/DateLib";
 import { convertFromUnits, convertUintToPercentage } from "@/utils/bn";
-import { formatCurrency } from "@/utils/formatter/currency";
 import { formatPercent } from "@/utils/formatter/percent";
 import { Trans } from "@lingui/macro";
+import { useNumberFormat } from "@/src/hooks/useNumberFormat";
 
 export const PolicyFeesAndExpiry = ({ data, coverPeriod }) => {
   const { fee, rate } = data;
+  const { formatCurrency } = useNumberFormat();
 
   const rateConverted = convertUintToPercentage(rate);
   const coverFee = convertFromUnits(fee).toString();
