+++ src/modules/my-policies/MyPoliciesTxsTable.jsx
@@ -14,9 +14,9 @@ import { useTokenSymbol } from "@/src/hooks/useTokenSymbol";
 import { useCoverInfo } from "@/src/hooks/useCoverInfo";
 import { fromNow } from "@/utils/formatter/relative-time";
 import DateLib from "@/lib/date/DateLib";
-import { formatCurrency } from "@/utils/formatter/currency";
 import { useNetwork } from "@/src/context/Network";
 import { t, Trans } from "@lingui/macro";
+import { useNumberFormat } from "@/src/hooks/useNumberFormat";
 
 const renderHeader = (col) => (
   <th
@@ -121,6 +121,7 @@ export const MyPoliciesTxsTable = () => {
 
 const DetailsRenderer = ({ row }) => {
   const { coverInfo } = useCoverInfo(row.cover.id);
+  const { formatCurrency } = useNumberFormat();
 
   return (
     <td className="px-6 py-6">
@@ -147,6 +148,7 @@ const DetailsRenderer = ({ row }) => {
 const CxDaiAmountRenderer = ({ row }) => {
   const { register } = useRegisterToken();
   const tokenSymbol = useTokenSymbol(row.cxToken);
+  const { formatCurrency } = useNumberFormat();
 
   return (
     <td className="px-6 py-6 text-right">
