+++ src/modules/pools/pod-staking/PodStakingCard.jsx
@@ -19,13 +19,13 @@ import { useTokenSymbol } from "@/src/hooks/useTokenSymbol";
 import { config } from "@neptunemutual/sdk";
 import { useNetwork } from "@/src/context/Network";
 import { explainInterval } from "@/utils/formatter/interval";
-import { formatCurrency } from "@/utils/formatter/currency";
 import { useTokenName } from "@/src/hooks/useTokenName";
 import { Badge } from "@/common/Badge/Badge";
 import { formatPercent } from "@/utils/formatter/percent";
 import { PoolTypes } from "@/src/config/constants";
 import { getApr } from "@/src/services/protocol/staking-pool/info/apr";
 import { t, Trans } from "@lingui/macro";
+import { useNumberFormat } from "@/src/hooks/useNumberFormat";
 
 // data from subgraph
 // info from `getInfo` on smart contract
@@ -36,6 +36,7 @@ export const PodStakingCard = ({ data, tvl, getPriceByAddress }) => {
     key: data.key,
     type: PoolTypes.POD,
   });
+  const { formatCurrency } = useNumberFormat();
 
   const rewardTokenAddress = info.rewardToken;
   const stakingTokenSymbol = useTokenSymbol(info.stakingToken);
