+++ src/modules/reporting/active/ActiveReportingCard.jsx
@@ -3,7 +3,6 @@ import { OutlinedCard } from "@/common/OutlinedCard/OutlinedCard";
 import { ProgressBar } from "@/common/ProgressBar/ProgressBar";
 import { getCoverImgSrc } from "@/src/helpers/cover";
 import { useCoverInfo } from "@/src/hooks/useCoverInfo";
-import { formatCurrency } from "@/utils/formatter/currency";
 import { fromNow } from "@/utils/formatter/relative-time";
 import DateLib from "@/lib/date/DateLib";
 import { formatPercent } from "@/utils/formatter/percent";
@@ -14,12 +13,14 @@ import { CardStatusBadge } from "@/common/CardStatusBadge";
 import { useMyLiquidityInfo } from "@/src/hooks/provide-liquidity/useMyLiquidityInfo";
 import { useCommitment } from "@/src/hooks/provide-liquidity/useCommitment";
 import { Trans } from "@lingui/macro";
+import { useNumberFormat } from "@/src/hooks/useNumberFormat";
 
 export const ActiveReportingCard = ({ coverKey, incidentDate }) => {
   const { coverInfo } = useCoverInfo(coverKey);
   const statusInfo = useCoverStatusInfo(coverKey);
   const { info: liquidityInfo } = useMyLiquidityInfo({ coverKey });
   const { commitment } = useCommitment({ coverKey });
+  const { formatCurrency } = useNumberFormat();
 
   const imgSrc = getCoverImgSrc({ key: coverKey });
