+++ src/modules/reporting/resolved/ResolvedReportSummary.jsx
@@ -7,17 +7,18 @@ import { convertFromUnits, isGreater } from "@/utils/bn";
 import BigNumber from "bignumber.js";
 import { truncateAddress } from "@/utils/address";
 import { useFinalizeIncident } from "@/src/hooks/useFinalizeIncident";
-import { formatCurrency } from "@/utils/formatter/currency";
 import DateLib from "@/lib/date/DateLib";
 import { VotesSummaryHorizontalChart } from "@/src/modules/reporting/VotesSummaryHorizontalChart";
 import { formatPercent } from "@/utils/formatter/percent";
 import { t, Trans } from "@lingui/macro";
+import { useNumberFormat } from "@/src/hooks/useNumberFormat";
 
 export const ResolvedReportSummary = ({ incidentReport, refetchReport }) => {
   const { finalize, finalizing } = useFinalizeIncident({
     coverKey: incidentReport.key,
     incidentDate: incidentReport.incidentDate,
   });
+  const { formatCurrency } = useNumberFormat();
 
   const votes = {
     yes: convertFromUnits(incidentReport.totalAttestedStake)
