+++ src/pages/my-policies/[cover_id]/[timestamp]/claim.jsx
@@ -10,12 +10,12 @@ import { useCoverInfo } from "@/src/hooks/useCoverInfo";
 import { convertFromUnits } from "@/utils/bn";
 import { toBytes32 } from "@/src/helpers/cover";
 import { useActivePoliciesByCover } from "@/src/hooks/useActivePoliciesByCover";
-import { formatCurrency } from "@/utils/formatter/currency";
 import { ComingSoon } from "@/common/ComingSoon";
 import { useFetchReportsByKeyAndDate } from "@/src/hooks/useFetchReportsByKeyAndDate";
 import { Alert } from "@/common/Alert/Alert";
 import { isFeatureEnabled } from "@/src/config/environment";
 import { t, Trans } from "@lingui/macro";
+import { useNumberFormat } from "@/src/hooks/useNumberFormat";
 
 export function getServerSideProps() {
   return {
@@ -36,6 +36,7 @@ export default function ClaimPolicy({ disabled }) {
       coverKey,
       incidentDate: timestamp,
     });
+  const { formatCurrency } = useNumberFormat();
 
   const title = coverInfo?.projectName;
